add classification dir and move files
