+++ classification/start_cometml_albumentations.py
@@ -30,9 +30,9 @@ import torch.utils.data as data
 import torchvision
 from torchvision import datasets, models, transforms
 
-from dataset import MyDataset, make_filepath_list
-from evaluator import plot_roc_fig
-from train import train_model_cometml, visualize_model_cometml
+from classification.dataset import MyDataset, make_filepath_list
+from classification.evaluator import plot_roc_fig
+from classification.train import train_model_cometml, visualize_model_cometml
 
 # %%
 # loggingを開始 -> API情報を載せた'.comet.config'をhomeディレクトリに作成しているので、APIの入力は必要ない
