+++ classification/train.py
@@ -16,8 +16,8 @@ from sklearn.metrics import *
 import torch
 from torch.utils.tensorboard import SummaryWriter
 
-from dataset import split_dataset_into_3
-from evaluator import *
+from classification.dataset import split_dataset_into_3
+from classification.evaluator import *
 
 # 学習ループ
 def train_simple_model(model, dataloaders, class_names, device, criterion, optimizer, scheduler, num_epochs=25):
